+++ src/main/java/roomescape/auth/jwt/JwtTokenProvider.java
@@ -11,8 +11,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import roomescape.auth.domain.Role;
 import roomescape.auth.dto.LoginMember;
-import roomescape.exception.ExpiredTokenException;
-import roomescape.exception.UnauthenticatedUserException;
+import roomescape.common.exception.ExpiredTokenException;
+import roomescape.common.exception.UnauthenticatedUserException;
 import roomescape.member.domain.Member;
 
 @Component
