+++ src/test/java/roomescape/common/util/CookieUtilsTest.java
@@ -1,4 +1,4 @@
-package roomescape.util;
+package roomescape.common.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
