+++ src/test/java/roomescape/common/validator/DateRangeValidatorTest.java
@@ -1,4 +1,4 @@
-package roomescape.validator;
+package roomescape.common.validator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
