+++ src/test/java/roomescape/member/controller/MemberApiControllerTest.java
@@ -6,9 +6,9 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.config.IntegrationTest;
+import roomescape.common.config.IntegrationTest;
 import roomescape.member.dto.MemberSignUpRequest;
-import roomescape.util.CookieUtils;
+import roomescape.common.util.CookieUtils;
 
 class MemberApiControllerTest extends IntegrationTest {
