+++ src/test/java/roomescape/reservation/controller/ReservationTimeApiControllerTest.java
@@ -12,9 +12,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import roomescape.config.IntegrationTest;
+import roomescape.common.config.IntegrationTest;
 import roomescape.reservation.dto.request.TimeSaveRequest;
-import roomescape.util.CookieUtils;
+import roomescape.common.util.CookieUtils;
 
 class ReservationTimeApiControllerTest extends IntegrationTest {
