+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -5,18 +5,18 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static roomescape.reservation.domain.Status.SUCCESS;
 import static roomescape.reservation.domain.Status.WAIT;
-import static roomescape.util.Fixture.HORROR_DESCRIPTION;
-import static roomescape.util.Fixture.HORROR_THEME;
-import static roomescape.util.Fixture.HORROR_THEME_NAME;
-import static roomescape.util.Fixture.JOJO_EMAIL;
-import static roomescape.util.Fixture.JOJO_NAME;
-import static roomescape.util.Fixture.LOGIN_JOJO;
-import static roomescape.util.Fixture.MEMBER_JOJO;
-import static roomescape.util.Fixture.MEMBER_KAKI;
-import static roomescape.util.Fixture.RESERVATION_TIME_10_00;
-import static roomescape.util.Fixture.THUMBNAIL;
-import static roomescape.util.Fixture.TODAY;
-import static roomescape.util.Fixture.TOMORROW;
+import static roomescape.Fixture.HORROR_DESCRIPTION;
+import static roomescape.Fixture.HORROR_THEME;
+import static roomescape.Fixture.HORROR_THEME_NAME;
+import static roomescape.Fixture.JOJO_EMAIL;
+import static roomescape.Fixture.JOJO_NAME;
+import static roomescape.Fixture.LOGIN_JOJO;
+import static roomescape.Fixture.MEMBER_JOJO;
+import static roomescape.Fixture.MEMBER_KAKI;
+import static roomescape.Fixture.RESERVATION_TIME_10_00;
+import static roomescape.Fixture.THUMBNAIL;
+import static roomescape.Fixture.TODAY;
+import static roomescape.Fixture.TOMORROW;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -28,7 +28,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import roomescape.auth.domain.Role;
 import roomescape.auth.dto.LoginMember;
-import roomescape.config.DatabaseCleaner;
+import roomescape.common.config.DatabaseCleaner;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.domain.Description;
